Suman/forget password

+++ controllers/resetpassword.controller.js
@@ -2,17 +2,20 @@ require("dotenv").config();
 const argon2 = require("argon2");
 const { PASSWORDRESET } = require("../model/resetPassword.model");
 const { USER } = require("../module/users/user.model");
+const { sendSmsToRecipient } = require("../services/sendOtpToNumber");
 const { default: axios } = require("axios");
 
 // Reset password if user forget his password
+
 const userExists = async (req, res) => {
   const { emailId } = req.body;
+
   try {
     let user = await USER.findOne({ emailId });
     if (!user) {
       return res.status(404).json({ message: "User not found" });
     }
-
+    // otp to mail
     const otp = (Math.floor(Math.random() * 9000) + 1000).toString();
     const html = `
       <h4>Hi ${user.name},</h4>
@@ -27,40 +30,67 @@ const userExists = async (req, res) => {
         name: process.env.BREVO_SENDER_NAME,
         email: process.env.BREVO_SENDER_EMAIL,
       },
-      to: [{ email: emailId, name: "User" }],
+      to: [{ email: emailId, name: user.name || "User" }],
       subject: "Your Password Reset OTP",
       htmlContent: html,
     };
 
-    await axios.post("https://api.brevo.com/v3/smtp/email", emailData, {
-      headers: {
-        accept: "application/json",
-        "api-key": process.env.BREVO_SENDER_API_KEY,
-        "content-type": "application/json",
-      },
-    });
+    try {
+      await axios.post("https://api.brevo.com/v3/smtp/email", emailData, {
+        headers: {
+          accept: "application/json",
+          "api-key": process.env.BREVO_SENDER_API_KEY,
+          "content-type": "application/json",
+        },
+      });
+    } catch (emailError) {
+      console.error("Error sending email:", emailError);
+      return res.status(500).json({ message: "Failed to send OTP via email." });
+    }
 
-    await PASSWORDRESET.deleteMany({ emailId });
-    await PASSWORDRESET.create({ emailId, otp });
+    // OTP to SMS
+    const { contactNo } = user;
+    if (!contactNo) {
+      return res
+        .status(400)
+        .json({ message: "User contact number is missing." });
+    }
 
-    return res.status(200).json({ message: "OTP email sent successfully" });
+    const otpSMS = (Math.floor(Math.random() * 9000) + 1000).toString();
+    const text = `Your OTP for Sandhee Platform is ${otpSMS}. It is valid for 5 minutes. Please do not share it with anyone. Team SANDHEE (RecQARZ)`;
+
+    try {
+      sendSmsToRecipient(contactNo, text);
+      // Save OTPs in database
+      await PASSWORDRESET.deleteMany({ emailId });
+      await PASSWORDRESET.create({ emailId, otp, otpSMS });
+      return res
+        .status(200)
+        .json({ message: "OTP sent successfully to Number and Email" });
+    } catch (smsError) {
+      console.error("Error sending SMS:", smsError);
+      return res.status(500).json({ message: "Failed to send OTP via SMS." });
+    }
   } catch (err) {
     console.error("Error in userExists function:", err);
-    return res
-      .status(500)
-      .json({ message: err.message || "Internal Server Error" });
+    return res.status(500).json({ message: "Internal Server Error" });
   }
 };
 
+// After putting otp
 const verifyOtp = async (req, res) => {
-  const { emailId, otp } = req.body;
+  const { emailId, otp, otpSMS } = req.body;
   try {
     const userPassword = await PASSWORDRESET.findOne({ emailId: emailId });
     if (!userPassword) {
       return res.status(404).json({ message: "User not found" });
     }
     if (userPassword.otp !== otp) {
-      return res.status(401).json({ message: "Invalid OTP" });
+      return res.status(401).json({ message: "Invalid Email OTP" });
+    }
+
+    if (userPassword.otpSMS !== otpSMS) {
+      return res.status(401).json({ message: "Invalid Number OTP" });
     }
     await PASSWORDRESET.deleteMany({ emailId: emailId });
     return res.status(200).json({ message: "OTP verified successfully" });
@@ -69,6 +99,7 @@ const verifyOtp = async (req, res) => {
   }
 };
 
+//update password after putting right otp
 const updatepassword = async (req, res) => {
   const { emailId, password } = req.body;
   try {
